# code/backend/src/config/swagger.yaml
openapi: 3.0.0
info:
  title: API Reside
  version: 1.0.0
  description: API do Sistema de Gerenciamento de Condomínios Reside
servers:
  - url: https://reside-production.up.railway.app
    description: Servidor de Desenvolvimento Local
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    ErrorResponse:
      type: object
      properties:
        message:
          type: string
          description: Uma mensagem descritiva do erro.
          example: 'Token inválido ou expirado.'
...
    ResidentInfo:
      type: object
      properties:
        id:
          type: integer
          example: 42
        first_name:
          type: string
          example: "João"
        last_name:
          type: string
          example: "Silva"
        email:
          type: string
          format: email
          example: "joao.silva@email.com"
        apartment:
          type: string
          example: "201B"
        condominium_id:
          type: integer
          example: 1
        role:
          type: string
          enum: [user, admin, syndic]
          example: "user"
        contact_phone:
          type: string
          example: "11987654321"
    ResidentInfoResponse:
      type: object
      properties:
        resident:
          $ref: '#/components/schemas/ResidentInfo'
paths:
...
  /resident/me:
    get:
      tags:
        - Residentes
      summary: Obter informações do residente autenticado
      description: Retorna dados do residente com base no token Firebase do usuário logado
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Informações do residente retornadas com sucesso
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResidentInfoResponse'
        '401':
          description: Token não fornecido ou inválido
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
                example:
                  message: 'Token de autenticação não fornecido'
        '404':
          description: Residente não encontrado para o usuário autenticado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
                example:
                  message: 'Residente não encontrado'
        '500':
          description: Erro interno do servidor
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
                example:
                  message: 'Erro ao buscar informações do residente'
...